+++ MainGame/defaults.hpp
@@ -2,10 +2,10 @@
 
 #include <cstdint>
 
-constexpr uintmax_t ScreenWidth = 768;
-constexpr uintmax_t ScreenHeight = 576;
+constexpr size_t ScreenWidth = 768;
+constexpr size_t ScreenHeight = 576;
 
-constexpr uintmax_t PlayfieldWidth = 576;
-constexpr uintmax_t PlayfieldHeight = 576;
+constexpr size_t PlayfieldWidth = 576;
+constexpr size_t PlayfieldHeight = 576;
 
-constexpr uintmax_t DefaultTileSize = 32;
+constexpr size_t DefaultTileSize = 32;
